[OV-10] refactor(ai): Service 분리(#21)
